requirements.txt
Django==4.2.9
django-cors-headers==3.13.0
django-debug-toolbar==4.4.6
django-filter==22.1
django_jazzmin==2.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
gunicorn==22.0.0
Pillow==9.2.0
psycopg2-binary==2.9.9
pytest-django==4.9.0
pytest-factoryboy==2.7.0
python-dotenv==1.0.0
pytz==2022.4
redis==5.0.4
setuptools==75.1.0